requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dataclasses==0.6
Django==5.0.6
django-extensions==3.2.3
djangorestframework==3.15.2
djongo==1.3.6
dnspython==2.6.1
flower==2.0.1
humanize==4.9.0
idna==3.7
kombu==5.3.7
pika==1.3.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
pymongo==4.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
redis==5.0.7
redis-cli==1.0.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
tornado==6.4.1
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
